# docs/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Token API
  version: 1.0.0
  description: Версионированный API платформы для репетиторов и учеников.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /auth/register:
    post:
      security: []
      tags: [Identity]
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201': { $ref: '#/components/responses/CommandSuccess' }
        '422': { $ref: '#/components/responses/Problem' }
  /auth/login:
    post:
      security: []
      tags: [Identity]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { $ref: '#/components/responses/CommandSuccess' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /auth/login/code:
    post:
      security: []
      tags: [Identity]
      operationId: completeLoginWithCode
      description: >-
        Второй шаг входа. Challenge выдаётся за верный пароль, код доказывает
        владение почтой. Успех запоминает устройство на 30 дней.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginCodeRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/CommandSuccess' }
        '400': { $ref: '#/components/responses/Problem' }
  /auth/login/code/resend:
    post:
      security: []
      tags: [Identity]
      operationId: resendLoginCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge]
              properties:
                challenge: { type: string, minLength: 20, maxLength: 500 }
              additionalProperties: false
      responses:
        '202': { $ref: '#/components/responses/CommandSuccess' }
        '400': { $ref: '#/components/responses/Problem' }
  /auth/email/verify-code:
    post:
      security: []
      tags: [Identity]
      operationId: verifyEmailWithCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email }
                code: { type: string, minLength: 9, maxLength: 40 }
              additionalProperties: false
      responses:
        '200': { $ref: '#/components/responses/CommandSuccess' }
        '400': { $ref: '#/components/responses/Problem' }
  /auth/email/resend:
    post:
      security: []
      tags: [Identity]
      operationId: resendEmailVerification
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/EmailRequest' } } }
      responses:
        '202': { $ref: '#/components/responses/CommandSuccess' }
...
  /auth/password/forgot:
    post:
      security: []
      tags: [Identity]
      operationId: requestPasswordReset
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/EmailRequest' } } }
      responses:
        '202': { $ref: '#/components/responses/CommandSuccess' }
  /auth/password/reset:
    post:
      security: []
      tags: [Identity]
      operationId: resetPassword
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ResetPasswordRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/CommandSuccess' }
        '400': { $ref: '#/components/responses/Problem' }
...
  /invites:
    get:
      tags: [Relationships]
      operationId: listInvites
      parameters: *listParameters
      responses: { '200': { $ref: '#/components/responses/Paginated' } }
    post:
      tags: [Relationships]
      operationId: createInvite
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateInviteRequest' } } }
      responses:
        '200': { $ref: '#/components/responses/CommandSuccess' }
        '409': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: token_sid
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 8, maxLength: 200 }
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, instance, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }
        requestId: { type: string }
        errors: { type: array, items: { type: object } }
    RegisterRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name: { type: string, minLength: 2, maxLength: 200 }
        email: { type: string, format: email }
        password: { type: string, minLength: 10, maxLength: 200 }
        role: { type: string, enum: [student, tutor] }
        grade: { type: integer, minimum: 1, maximum: 11 }
        school: { type: string, maxLength: 300 }
        subjects: { type: array, items: { type: string }, maxItems: 20 }
        yearsExp: { type: number, minimum: 0, maximum: 80 }
        rate: { type: number, minimum: 0 }
        meetingUrl: { type: string, maxLength: 2000 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1, maxLength: 200 }
    LoginCodeRequest:
      type: object
      required: [challenge, code]
      properties:
        challenge: { type: string, minLength: 20, maxLength: 500 }
        code:
          type: string
          minLength: 9
          maxLength: 40
          description: Код из письма. Регистр и разделители произвольны.
          example: K7M-2PQ-9XZ
    EmailRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token: { type: string, minLength: 20, maxLength: 500 }
        password: { type: string, minLength: 10, maxLength: 200 }
...
    CreateInviteRequest:
      type: object
      required: [kind]
      properties:
        kind: { type: string, enum: [enrollment, group] }
        subjectId: { type: string }
        groupId: { type: string }
        maxUses: { type: [integer, 'null'], minimum: 1 }
        expiresAt: { type: [string, 'null'], format: date-time }
        note: { type: string, maxLength: 2000 }
...
    PageInfo:
      type: object
      required: [limit, hasMore, nextCursor]
      properties:
        limit: { type: integer }
        hasMore: { type: boolean }
        nextCursor: { type: [string, 'null'] }
    Paginated:
      type: object
      required: [items, page, meta]
      properties:
        items: { type: array, items: {} }
        page: { $ref: '#/components/schemas/PageInfo' }
        meta: { type: object, additionalProperties: true }
  responses:
    Problem:
      description: Ошибка в формате RFC 9457 Problem Details
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    CommandSuccess:
      description: Команда выполнена
      content: { application/json: { schema: { type: object, additionalProperties: true } } }
...
    Paginated:
      description: Страница ресурсов
      headers: { ETag: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Paginated' } } }